manual-implementation/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
Deprecated==1.2.13
Django==4.0.4
django-redis==5.2.0
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
ifaddr==0.1.7
packaging==21.3
py-eureka-client==0.10.5
pymongo==3.12.3
pyparsing==3.0.9
pytz==2022.1
redis==4.3.1
sqlparse==0.2.4
tzdata==2022.1
vine==5.0.0
wrapt==1.14.1
